extract `testAcrossTimezones`
